rename from sct.register to sct.registration
