+++ spinalcordtoolbox/registration/landmarks.py
